refactor: refactor main preview component
