+++ apps/studio/src/components/render/index.ts
@@ -1,0 +1,3 @@
+export { createRenderBlock } from './renderBlock';
+export { renderInline } from './renderInline';
+export { createRenderNode } from './renderNode';
